new/requirements.txt
google-genai
tqdm
flask
flask-sqlalchemy
apscheduler
datetime
requests
PySocks>=1.7.1 